Add an IE9 compatible ajax method using XDRs
